Margarida's_Garden_Project/requirements.txt
# Margarida's Garden - Dependências

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Templates
jinja2==3.1.3
email-validator==2.1.0

# Banco de Dados (SQLAlchemy >= 2.0.41 para Python 3.14)
sqlalchemy>=2.0.41
alembic>=1.13.1

# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9

# Criptografia de senhas armazenadas
cryptography==42.0.2

# Variáveis de ambiente
python-dotenv==1.0.1
pydantic-settings>=2.0.0

# Rate limiting
slowapi>=0.1.9

# Testes
pytest>=7.0.0,<8
pytest-asyncio>=0.23.0
httpx==0.26.0
